Fix: Ask for passphrase if pulumi uses local
When using Pulumi with local service, a passphrase needs
to be generated and used for each command.
Since the pulumi CLI is executed in non interactive mode,
we need to pass the passphrase as env variable.
As first step, the code is checking if Pulumi is running with local
(using `pulumi login` output). If it does, it prompts the user
to provide the passphrase and add that to the env variables

Fix #1
Fix #2